LSTM-CNN_Fusion/requirements.txt
torch>=1.3.0 
numpy>=1.15.0 
pandas>=0.23.0 
biopython>=1.70